create preflight and commit_hook modules
